Added spring security code authentication
