+++ src/main/java/com/journal/journalApp/controller/PublicController.java
@@ -1,0 +1,30 @@
+package com.journal.journalApp.controller;
+
+import com.journal.journalApp.entity.User;
+import com.journal.journalApp.repository.UserRepository;
+import com.journal.journalApp.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/public")
+public class PublicController {
+
+    @Autowired
+    private UserService userService;
+
+    @PostMapping("/create-user")
+    public ResponseEntity<ResponseEntity> createUser(@RequestBody User user){
+        try {
+            userService.saveEntry(user);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+    }
+}
