+++ src/main/java/com/journal/journalApp/controller/UserController.java
@@ -6,6 +6,8 @@ import com.journal.journalApp.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,29 +18,16 @@ public class UserController {
 
     @Autowired
     private UserService userService;
-
-   /* @PostMapping
-    public void createUser(@RequestBody User user){
-        userService.saveEntry(user);
-    }*/
-    @PostMapping
-    public ResponseEntity<ResponseEntity> createUser(@RequestBody User user){
-        try {
-            userService.saveEntry(user);
-            return new ResponseEntity<>(HttpStatus.CREATED);
-        }catch (Exception e){
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
-    }
-
     @GetMapping
     public List<User> getAllUsers(){
         return userService.findAll();
     }
 
-    @PutMapping("/{userName}")
-    public ResponseEntity<?> updateUser(@RequestBody User user, @PathVariable String userName){
-        User userInDB = userService.findByUserName(userName);
+    @PutMapping
+    public ResponseEntity<?> updateUser(@RequestBody User user){
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User userInDB = userService.findByUserName(authentication.getName());
         if(userInDB != null){
             userInDB.setUserName(user.getUserName());
             userInDB.setPassword(user.getPassword());
