+++ src/main/java/com/journal/journalApp/service/UserService.java
@@ -4,8 +4,11 @@ import com.journal.journalApp.entity.User;
 import com.journal.journalApp.repository.UserRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,8 +18,11 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
+    private static final PasswordEncoder password = new BCryptPasswordEncoder();
     //save
     public void saveEntry(User user){
+        user.setPassword(password.encode(user.getPassword()));
+        user.setRoles(Arrays.asList("USER"));
         userRepository.save(user);
     }
