+++ src/main/resources/application.properties
@@ -3,4 +3,3 @@ spring.application.name=journalApp
 spring.data.mongodb.uri=mongodb+srv://riteshkothalemongodb:MIzST8FMrAtBdNZM@cluster0.k98ha.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
 spring.data.mongodb.database=journaldb
 spring.data.mongodb.auto-index-creation=true
-
